Tugas 10 dpw

+++ system/app/Models/Traits/Attributes/ProdukAttributes.php
@@ -1,13 +1,10 @@
 <?php
+
 namespace App\Models\Traits\Attributes;
 
 use Illuminate\Support\Str;
 
-trait ProdukAttributes{
-    
-    function getHargaStringAttribute(){
-        return "Rp.".number_format($this->attributes['harga']);
-    }
+trait ProdukAttributes {
 
     function handleUploadFoto(){
         $this->handleDelete();
@@ -17,19 +14,21 @@ trait ProdukAttributes{
             $randomStr = Str::random(5);
             $filename = $this->id."-".time()."-".$randomStr.".".$foto->extension();
             $url = $foto->storeAs($destination, $filename);
-            $this->foto = "app/".$url; 
+            $this->foto = "app/".$url;
             $this->save();
-           
         }
     }
 
     function handleDelete(){
-       $foto = $this->foto;
-       $path = public_path($foto);
-       if(file_exists($path)){
-        unlink($path);
-       }
-       return true;
-    }
+        $foto = $this->foto;
+        if($foto){
+        $path = public_path($foto);
+            if(file_exists($path)){
+                unlink($path);
+            }
+        }
+        return true;
 
+    }
+    
 }
